Remove dependency from json-stringify-safe
We don't need this feature in the nexa explorer and we need to remove
it so that jayson will use JSON.stringify directly. This will let us
use an experimental version of v8 engine that will probably fix
JSON BigInt parsing/stringifying problem.

See the followings links for more detail:

- https://jsoneditoronline.org/indepth/parse/why-does-json-parse-corrupt-large-numbers/
- https://github.com/tc39/proposal-json-parse-with-source
- https://2ality.com/2022/11/json-parse-with-source.html

The TC39 change proposal is called json parse with souce and it has already been
implemented in google v8 since version 10.9.1, see

https://chromium.googlesource.com/v8/v8/+/refs/heads/10.9.1/src/flags/flag-definitions.h#222

Nodejs ver >= 20 need to be used cause those are the only once that come with an
embedded v8 with ver >= 10.9.1, eg

```
$> node -p process.versions.v8
11.3.244.8-node.10
$> node --version
v20.5.1
```